format everything and add intellij project files

+++ src/client/command/commands/gm6/ShutdownCommand.java
@@ -23,9 +23,9 @@
 */
 package client.command.commands.gm6;
 
-import client.command.Command;
-import client.MapleClient;
 import client.MapleCharacter;
+import client.MapleClient;
+import client.command.Command;
 import net.server.Server;
 import net.server.world.World;
 import server.TimerManager;
@@ -38,23 +38,23 @@ public class ShutdownCommand extends Command {
     @Override
     public void execute(MapleClient c, String[] params) {
         MapleCharacter player = c.getPlayer();
-        if (params.length < 1){
+        if (params.length < 1) {
             player.yellowMessage("Syntax: !shutdown [<time>|NOW]");
             return;
         }
-        
+
         int time = 60000;
-        if (params[0].equalsIgnoreCase("now")){
+        if (params[0].equalsIgnoreCase("now")) {
             time = 1;
         } else {
             time *= Integer.parseInt(params[0]);
         }
 
         if (time > 1) {
-            int seconds = (int) (time / 1000) % 60;
-            int minutes = (int) ((time / (1000 * 60)) % 60);
-            int hours = (int) ((time / (1000 * 60 * 60)) % 24);
-            int days = (int) ((time / (1000 * 60 * 60 * 24)));
+            int seconds = (time / 1000) % 60;
+            int minutes = (time / (1000 * 60)) % 60;
+            int hours = (time / (1000 * 60 * 60)) % 24;
+            int days = (time / (1000 * 60 * 60 * 24));
 
             String strTime = "";
             if (days > 0) strTime += days + " days, ";
@@ -68,7 +68,7 @@ public class ShutdownCommand extends Command {
                 }
             }
         }
-        
+
         TimerManager.getInstance().schedule(Server.getInstance().shutdown(false), time);
     }
 }
