+++ src/net/server/channel/handlers/RaiseIncExpHandler.java
@@ -1,22 +1,21 @@
 package net.server.channel.handlers;
 
-import java.util.Map;
-
 import client.MapleCharacter;
 import client.MapleClient;
 import client.MapleQuestStatus;
 import client.inventory.MapleInventory;
 import client.inventory.MapleInventoryType;
-import net.AbstractMaplePacketHandler;
 import client.inventory.manipulator.MapleInventoryManipulator;
+import net.AbstractMaplePacketHandler;
 import server.MapleItemInformationProvider;
 import server.MapleItemInformationProvider.QuestConsItem;
 import server.quest.MapleQuest;
 import tools.MaplePacketCreator;
 import tools.data.input.SeekableLittleEndianAccessor;
 
+import java.util.Map;
+
 /**
- *
  * @author Xari
  * @author Ronan - added concurrency protection and quest progress limit
  */
@@ -27,7 +26,7 @@ public class RaiseIncExpHandler extends AbstractMaplePacketHandler {
         byte inventorytype = slea.readByte();//nItemIT
         short slot = slea.readShort();//nSlotPosition
         int itemid = slea.readInt();//nItemID
-        
+
         if (c.tryacquireClient()) {
             try {
                 MapleItemInformationProvider ii = MapleItemInformationProvider.getInstance();
@@ -38,14 +37,14 @@ public class RaiseIncExpHandler extends AbstractMaplePacketHandler {
 
                 int infoNumber = consItem.questid;
                 Map<Integer, Integer> consumables = consItem.items;
-                
+
                 MapleCharacter chr = c.getPlayer();
                 MapleQuest quest = MapleQuest.getInstanceFromInfoNumber(infoNumber);
                 if (!chr.getQuest(quest).getStatus().equals(MapleQuestStatus.Status.STARTED)) {
                     c.announce(MaplePacketCreator.enableActions());
                     return;
                 }
-                
+
                 int consId;
                 MapleInventory inv = chr.getInventory(MapleInventoryType.getByType(inventorytype));
                 inv.lockInventory();
@@ -55,15 +54,15 @@ public class RaiseIncExpHandler extends AbstractMaplePacketHandler {
                         return;
                     }
 
-                    MapleInventoryManipulator.removeFromSlot(c, MapleInventoryType.getByType(inventorytype), (short) slot, (short) 1, false, true);
+                    MapleInventoryManipulator.removeFromSlot(c, MapleInventoryType.getByType(inventorytype), slot, (short) 1, false, true);
                 } finally {
                     inv.unlockInventory();
                 }
-                
+
                 int questid = quest.getId();
                 int nextValue = Math.min(consumables.get(consId) + c.getAbstractPlayerInteraction().getQuestProgressInt(questid, infoNumber), consItem.exp * consItem.grade);
                 c.getAbstractPlayerInteraction().setQuestProgress(questid, infoNumber, nextValue);
-                
+
                 c.announce(MaplePacketCreator.enableActions());
             } finally {
                 c.releaseClient();
