+++ src/net/server/channel/handlers/SkillBookHandler.java
@@ -28,12 +28,13 @@ import client.SkillFactory;
 import client.inventory.Item;
 import client.inventory.MapleInventory;
 import client.inventory.MapleInventoryType;
-import java.util.Map;
-import net.AbstractMaplePacketHandler;
 import client.inventory.manipulator.MapleInventoryManipulator;
+import net.AbstractMaplePacketHandler;
 import server.MapleItemInformationProvider;
 import tools.MaplePacketCreator;
 import tools.data.input.SeekableLittleEndianAccessor;
+
+import java.util.Map;
 
 public final class SkillBookHandler extends AbstractMaplePacketHandler {
     @Override
@@ -42,16 +43,16 @@ public final class SkillBookHandler extends AbstractMaplePacketHandler {
             c.announce(MaplePacketCreator.enableActions());
             return;
         }
-        
+
         slea.readInt();
-        short slot = (short) slea.readShort();
+        short slot = slea.readShort();
         int itemId = slea.readInt();
-        
+
         boolean canuse;
         boolean success = false;
         int skill = 0;
         int maxlevel = 0;
-        
+
         MapleCharacter player = c.getPlayer();
         if (c.tryacquireClient()) {
             try {
@@ -94,7 +95,7 @@ public final class SkillBookHandler extends AbstractMaplePacketHandler {
             } finally {
                 c.releaseClient();
             }
-            
+
             // thanks Vcoc for noting skill book result not showing for all in area
             player.getMap().broadcastMessage(MaplePacketCreator.skillBookResult(player, skill, maxlevel, canuse, success));
         }
